Identify and ignore certs returned by the ssh-agent
Previously, anything returned in a SSH_AGENT_IDENTITIES_ANSWER message
was assumed to be parseable as a PublicKey by the ssh-key crate,
which is no the case for certs. Instead we now detect items returned
in the identities answer with key-type values that contains the substring
-cert- and ignore those.

There is also a small refactoring included that ensures that we never
panic when attempting to read a length field from a buffer that is
shorter than 4 bytes.

Closes #39