Have Client::list_all_identities() return 'static lifetimed value
Since list_all_identities() takes a &mut self reference, previously
that meant that the mutable borrow would need to live for as long
as the returned values lived. Since the implementation of
list_all_identities() only returns the Cow::Owned variant and as
such don't even hold borrowed values, this was a bit frustrating.

It seems like the received wisdom is that if the value returned
has the 'static lifetime, this resolves the problem.